Create Processing
This code generates a QR code that, when scanned, directs the user to a specified URL. In this case, the URL is "https://exampleChiro-X.com/pay". The code also contains a function that processes a payment amount, "process_payment(payment_amount)" and performs a series of operations, such as exchanging a portion of the payment amount for ChiroCoins and sending the ChiroCoins to a customer's XUMM wallet. The code has the necessary components to implement a system for exchanging payments for ChiroCoins, but it will require additional development and integration with an actual payment processor and XUMM wallet API to be fully functional.